feat(tui): inline transcript diffs — capped at 10 lines, ctrl+o expands
A successful file-mutating tool result now shows its diff inline in the
transcript, GitHub-PR style through the shared crate::diff renderer:
path header rule, gutter line numbers, add/remove backgrounds with
syntax colors layered under, and a +added/-removed footer. Collapsed
shows at most INLINE_DIFF_CAP (10) styled lines plus a fold hint;
ctrl+o (the deck's existing per-entry expansion) reveals the full diff.

The model already recorded InlineDiffRef on ToolResult entries — the
renderer just never consumed it. entry_lines now resolves it against
SessionModel::files, freshness-gated: a later mutation of the same path
bumps FileState::changes past the recorded seq and the diff is hidden
rather than misattributed. Because that staling appends nothing to the
transcript, both render caches gain a file-mutation fingerprint term
(TranscriptCache.file_gen, SessionFold key) so a settled entry refolds.

Ratchet: render.rs's inline mod tests moved to src/render/tests.rs
(same medicine as a3c8420) — 2655 → 1489 lines, allowlist entry
REMOVED. Drive-by: tests/progress_gold.rs field_reassign_with_default
broke clippy -D warnings; fixed with struct-update init.
