+++ verifications/transcript-inline-diff.txt
@@ -1,0 +1,37 @@
+# Verification — inline transcript diffs (capped, syntax-highlighted, ctrl+o)
+
+Feature: a successful file-mutating tool result (write_file/edit_file/
+delete_file) now renders its diff inline in the transcript, GitHub-PR style
+via the shared `crate::diff` renderer — header rule with the path, gutter
+line numbers, add/remove backgrounds with syntax colors layered under, and
+a closing rule with the +added/-removed counts. Collapsed shows at most
+INLINE_DIFF_CAP (10) styled diff lines with a `⋯ +N more diff lines ·
+ctrl+o opens the full diff` fold hint; ctrl+o (the existing per-entry
+expansion in the deck's Session tab, ctrl+r expand-all in the classic
+shell) reveals the whole diff. Freshness-gated: a later mutation of the
+same path bumps `FileState::changes` past the recorded `InlineDiffRef.seq`
+and the diff is hidden rather than misattributed.
+
+Witness tests (all fail on the pre-change renderer, which dropped the
+`diff` field on the floor):
+
+  cargo test -p stella-tui
+  test render::tests::collapsed_tool_result_shows_a_capped_syntax_highlighted_inline_diff ... ok
+  test render::tests::expanded_tool_result_shows_the_full_inline_diff ... ok
+  test render::tests::a_stale_or_unresolvable_diff_ref_renders_no_inline_diff ... ok
+  test render::tests::ui_transcript_cache_invalidates_when_a_file_mutation_stales_an_inline_diff ... ok
+  test views::session::tests::a_later_mutation_to_the_same_path_stales_the_settled_inline_diff ... ok
+  test result: ok. 487 passed; 0 failed; 1 ignored   (lib)
+  test result: ok. 4 passed  (tests/deck_snapshot.rs)
+  test result: ok. 1 passed  (tests/progress_gold.rs)
+
+Gate pieces run locally:
+  cargo clippy -p stella-tui --all-targets -- -D warnings   → clean
+  cargo fmt -p stella-tui -- --check                        → clean
+  bash scripts/check-file-sizes.sh                          → OK (262 files)
+
+Ratchet: render.rs's inline `mod tests` moved to src/render/tests.rs
+(same medicine as stella-model in a3c8420) — render.rs 2655 → 1489 lines,
+under the 1500 new-file cap, so its allowlist entry (2392) is REMOVED
+rather than raised. Drive-by fix: tests/progress_gold.rs had a
+field_reassign_with_default clippy error that broke `-D warnings`.
